requirements.txt
asgiref==3.2.10
cffi==1.14.0
coverage==5.2
cryptography==2.9.2
Django==3.0.8
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
